requirements.txt
# Core graph + LLM
neo4j==5.28.1
langchain==0.3.25
langchain-neo4j==0.4.0
langchain-google-genai==2.1.4
langchain-community==0.3.23
langgraph==0.4.1
# Data
pandas==2.2.3
python-dotenv==1.1.0

# API server (optional)
fastapi==0.115.12
uvicorn==0.34.2

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
